+++ src/layout/AppLayout.jsx
@@ -13,6 +13,8 @@ export function AppLayout() {
     flowAppId: params.flowAppId,
     docId: params.docId,
     manualId: params.manualId,
+    sectionId: params.sectionId,
+    formId: params.formId,
     topicId: params.topicId,
     projectId: isReference ? params.projectId : undefined,
     stepProjectId: isStepByStep ? params.projectId : undefined,